% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deme.R
\name{read_deme}
\alias{read_deme}
\title{Load and Validate a Demes Model}
\usage{
read_deme(file, text)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for writing}

\item{text}{character string: if file is not supplied and this is, then data are read from the value of text via a text connection. Notice that a literal string can be used to include (small) data sets within R code.}
}
\value{
\code{out}
}
\description{
More description
}
\examples{
\dontrun{file_name <- "test_file.yaml"
a <- read_deme(file=file_name)}

yaml_string <- "time_units: generations\ndemes:\n  - name: a\n    epochs:\n    - start_size: 100"
b <- read_deme(text=yaml_string)
}
