% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_demes.R
\name{read_demes}
\alias{read_demes}
\title{Load and Validate a Demes Model}
\usage{
read_demes(file, text)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for writing}

\item{text}{character string: if file is not supplied and this is, then data are read from the value of text via a text connection. Notice that a literal string can be used to include (small) data sets within R code.}
}
\value{
A fully specified Demes model, a nested list. See the Demes specifications for more details https://github.com/popsim-consortium/demes-spec/.
}
\description{
\code{read_demes} reads a Demes model yaml file or string, validates the input and creates a fully specified R object from it
}
\examples{
\dontrun{
file_name <- "test_file.yaml"
a <- read_demes(file=file_name)}

yaml_string <- "time_units: generations\ndemes:\n  - name: a\n    epochs:\n    - start_size: 100"
b <- read_demes(text=yaml_string)
}
