% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_deme.R
\name{load_deme}
\alias{load_deme}
\title{Load and Validate a Demes Model}
\usage{
load_deme(file, text)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for writing}

\item{text}{character string: if file is not supplied and this is, then data are read from the value of text via a text connection. Notice that a literal string can be used to include (small) data sets within R code.}
}
\value{
\code{out}
}
\description{
More description
}
