% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_demes.R
\name{read_demes}
\alias{read_demes}
\title{Load and validate a Demes model}
\usage{
read_demes(file, text)
}
\arguments{
\item{file}{Either a path to a YAML file or a connection open for writing}

\item{text}{If a \code{file} is not provided, Demes YAML input is
read as a literal string given in the \code{text}}
}
\value{
A fully instantiated Demes model as a nested R list
}
\description{
This function reads a Demes model from a YAML file or a literal YAML string,
validates the model, and returns it as a fully instantiated R object.
}
\details{
See the Demes specification document for more details about the Demes format
<https://github.com/popsim-consortium/demes-spec/>.
}
\examples{
path <- system.file("extdata/yaml", "minimal_01.yaml", package = "demes")
demes1 <- read_demes(file = path)

yaml_string <- "time_units: generations\ndemes:\n  - name: a\n    epochs:\n    - start_size: 100"
demes2 <- read_demes(text = yaml_string)
}
